GW-80: Refactored landing image and background color
