+++ client/src/assets/landing-image/partners.jpeg
