+++ client/src/pages/events/Events.jsx
@@ -2,12 +2,12 @@ import React from 'react';
 import { Box } from '@mui/material';
 import MainContainer from 'shared/layout/MainContainer';
 import { Typography, ScrollButton, LandingImage, Footer } from 'shared/components/index';
-import eventLandingImage from 'assets/landing/event.jpg';
+import eventLandingImage from 'assets/landing-image/event.jpg';
 import EventContent from './components/EventContent';
 
 export default function Events() {
 	return (
-		<Box className='bg-[#FFFDF5]'>
+		<Box>
 			<LandingImage bgImage={eventLandingImage} text='Our Events' />
 			<MainContainer>
 				<Box className='w-4/5'>
