+++ client/src/pages/partners/Partners.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { Box, Divider } from '@mui/material';
 import { ScrollButton, LandingImage, Footer } from 'shared/components/index';
-import partnersLandingImg from 'assets/landing/partners.jpeg';
+import partnersLandingImg from 'assets/landing-image/partners.jpeg';
 import MainContainer from 'shared/layout/MainContainer';
 import { SponsorRegistration, SponsorTitle, SponsorList } from './components/index';
 
 export default function Partners() {
 	return (
-		<Box className='bg-[#FFFDF5]'>
+		<Box>
 			<LandingImage bgImage={partnersLandingImg} text='Our Partners' />
 			<MainContainer>
 				<SponsorTitle />
